refactor: move ticket animation into Ticket.js component
